requirements-ingestion.txt
# Minimal requirements for the data ingestion pipeline only
# Excludes: streamlit, fastapi, ragas, pytest, jupyter (not needed for ingestion)

# AWS S3 document loading
boto3

# Document parsing
beautifulsoup4==4.14.3
pypdf==6.9.0

# LangChain ecosystem (pinned to match project)
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.83
langchain-openai==0.3.35
langchain-text-splitters==0.3.11

# OpenAI embeddings
openai==2.28.0
tiktoken==0.12.0

# Pinecone vector database
pinecone==8.1.0

# Sparse vector generation (BM25)
rank-bm25==0.2.2
numpy==2.4.3

# Configuration & logging
python-dotenv==1.2.2
python-json-logger==4.0.0
